# assets/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo API",
    "version": "1.0.0",
    "description": "API for managing todos"
  },
  "paths": {
    "/auth/login": {
      "get": {
        "summary": "Initiate Spotify OAuth login",
        "operationId": "spotifyAuth",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "302": {
            "description": "Redirect to Spotify authorization page"
          }
        }
      }
    },
    "/auth/login/callback": {
      "get": {
        "summary": "Spotify OAuth callback",
        "operationId": "spotifyCallbackAuth",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "description": "Authorization code from Spotify",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "description": "Error from Spotify OAuth",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to success or failure page"
          }
        }
      }
    },
    "/auth/success": {
      "get": {
        "summary": "Login success endpoint",
        "operationId": "loginSuccess",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Successful authentication",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "accessToken": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/failure": {
      "get": {
        "summary": "Login failure endpoint",
        "operationId": "loginFailure",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "401": {
            "description": "Login failed"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "Logout user",
        "operationId": "logout",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          }
        }
      }
    },
    "/spotify/artists-followed": {
      "get": {
        "summary": "Get followed artists",
        "operationId": "getArtistsFollowed",
        "tags": [
          "Spotify"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of followed artists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artists": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "uri": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated with Spotify",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/spotify/play": {
      "put": {
        "summary": "Play a song",
        "operationId": "playSong",
        "tags": [
          "Spotify"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "description": "Track details to play",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trackUri": {
                    "type": "string",
                    "description": "Spotify track URI"
                  },
                  "deviceId": {
                    "type": "string",
                    "description": "Optional device ID"
                  }
                },
                "required": [
                  "trackUri"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Song started playing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Track URI required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated with Spotify",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/spotify/top-tracks": {
      "get": {
        "summary": "Get user's top tracks",
        "operationId": "getTopTracks",
        "tags": [
          "Spotify"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of tracks to return (default: 10)",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "timeRange",
            "in": "query",
            "required": false,
            "description": "Time range for top tracks",
            "schema": {
              "type": "string",
              "enum": [
                "short_term",
                "medium_term",
                "long_term"
              ],
              "default": "medium_term"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User's top tracks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tracks": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "artists": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "uri": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated with Spotify",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/spotify/play-top": {
      "put": {
        "summary": "Play a top song by index",
        "operationId": "playTopSong",
        "tags": [
          "Spotify"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "description": "Song index to play from top tracks",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "songIndex": {
                    "type": "integer",
                    "description": "Index of song to play (1-10)",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "deviceId": {
                    "type": "string",
                    "description": "Optional device ID"
                  }
                },
                "required": [
                  "songIndex"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Top song started playing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Song index required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated with Spotify",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/spotify/pause": {
      "put": {
        "summary": "Pause playback",
        "operationId": "pausePlayback",
        "tags": [
          "Spotify"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "description": "Optional device ID",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deviceId": {
                    "type": "string",
                    "description": "Optional device ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Playback paused",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated with Spotify",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/spotify/current-playback": {
      "get": {
        "summary": "Get current playback information",
        "operationId": "getCurrentPlayback",
        "tags": [
          "Spotify"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current playback information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "device": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        },
                        "volume_percent": {
                          "type": "integer"
                        }
                      }
                    },
                    "is_playing": {
                      "type": "boolean"
                    },
                    "item": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "artists": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "progress_ms": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated with Spotify",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/spotify/currently-playing": {
      "get": {
        "summary": "Get currently playing track",
        "operationId": "getCurrentlyPlaying",
        "tags": [
          "Spotify"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Currently playing track information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "item": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "artists": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "uri": {
                          "type": "string"
                        }
                      }
                    },
                    "is_playing": {
                      "type": "boolean"
                    },
                    "progress_ms": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated with Spotify",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          }
        }
      },
      "Artist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          }
        }
      },
      "Track": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Artist"
            }
          },
          "uri": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session-based authentication using cookies"
      },
      "spotifyOAuth": {
        "type": "oauth2",
        "description": "Spotify OAuth 2.0",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://accounts.spotify.com/authorize",
            "tokenUrl": "https://accounts.spotify.com/api/token",
            "scopes": {
              "user-read-private": "Read access to user profile",
              "user-read-email": "Read access to user email",
              "user-follow-read": "Read access to followed artists",
              "user-top-read": "Read access to top tracks and artists",
              "user-modify-playback-state": "Control playback",
              "user-read-playback-state": "Read current playback state",
              "user-read-currently-playing": "Read currently playing track"
            }
          }
        }
      }
    }
  }
}